+++ Tests/MockDal.cs
@@ -7,9 +7,9 @@ namespace Tests
 {
     internal class MockDal : IDataAccessLayer
     {
-        private readonly List<int> _primeNumbers = new List<int> { 2, 3, 5, 7, 11, 13, 17, 19, 23 };
+        private readonly List<long> _primeNumbers = new List<long> { 2, 3, 5, 7, 11, 13, 17, 19, 23 };
 
-        public bool? IsPrime(int numberToTest)
+        public bool? IsPrime(long numberToTest)
         {
             if (numberToTest > LargestTestedNumber)
             {
